Move vec to shlr/sdb
