+++ shlr/sdb/include/sdb/r_vec.h
