+++ test/unit/test_vec.c
@@ -1,5 +1,5 @@
 #include <r_util.h>
-#include <r_vec.h>
+#include <sdb/r_vec.h>
 #include "minunit.h"
 
 typedef struct {
